[NEAR-119] Refactor: Concert 스키마 분리
- concert를 stream에서 독립
